Extract marker schema behind compatibility facade
